refactor: model 분리
