+++ src/Model/Car.js
@@ -3,8 +3,6 @@ class Car {
   constructor(carName) {
     this.carName = carName;
     this.count = 0;
-    this.template = new Template();
-    this.carNameTemplate = this.template.carNameBox(this.carName);
   }
 
   move() {
@@ -16,16 +14,6 @@ class Car {
       return false;
     }
     return true;
-  }
-}
-
-class Template {
-  carNameBox(carName) {
-    return `
-      <div class="car-name-box">
-        <div class="car-name-box-text">${carName}</div>
-      </div>
-    `;
   }
 }
